+++ src/components/disponibilites/discord-poll-manager/constants.ts
@@ -1,0 +1,6 @@
+export const JSON_HEADERS = {
+  "Content-Type": "application/json",
+} as const;
+
+export const DEFAULT_CLOSE_MESSAGE =
+  "Les inscriptions sont terminées pour cette journée.\n\nSi vous voulez vous ajoutez, il faut envoyer un message à Joffrey en privé.";
